+++ test/datalevin/tx_group_test.clj
@@ -4,8 +4,7 @@
   (:import [datalevin.tx_group Group]
            [java.util.concurrent ConcurrentLinkedQueue CountDownLatch TimeUnit]
            [java.util.concurrent.atomic AtomicBoolean AtomicInteger]
-           [java.util.concurrent.locks ReentrantLock]
-           [org.eclipse.collections.impl.list.mutable FastList]))
+           [java.util.concurrent.locks ReentrantLock]))
 
 (def ^:dynamic *submitted-value* nil)
 
@@ -20,12 +19,95 @@
 (defn- queued [^Group g]
   (.size ^ConcurrentLinkedQueue (.-queue g)))
 
+(defn- batch-size [result]
+  ;; These test operations return scalars or vectors; only a batch returns
+  ;; the executor's object array.
+  (if (instance? (Class/forName "[Ljava.lang.Object;") result)
+    (alength ^objects result)
+    1))
+
+(deftest uncontended-writes-execute-directly-without-wrapping-the-operation
+  (let [g (group/create 8)
+        caller (Thread/currentThread)
+        context (Object.)]
+    (doseq [value [nil false :done (object-array [1 2])]]
+      (binding [*submitted-value* value]
+        (let [op (fn [tx]
+                   (is (identical? context tx))
+                   (is (identical? caller (Thread/currentThread)))
+                   *submitted-value*)
+              result (group/submit! g (fn [execute]
+                                       (is (identical? op execute))
+                                       (execute context))
+                                    op)]
+          (is (identical? value result)))))
+    (is (zero? (queued g)))))
+
+(deftest direct-failures-propagate-once-and-release-leadership
+  (let [g (group/create 8)]
+    (doseq [body-failure? [true false]]
+      (let [failure (ex-info "direct transaction failed" {})
+            calls (atom 0)
+            seen (atom nil)
+            op (fn [_]
+                 (swap! calls inc)
+                 (when body-failure? (throw failure)))
+            run-transaction (fn [execute]
+                              (try
+                                (execute nil)
+                                (throw failure)
+                                (catch Throwable t
+                                  (reset! seen t)
+                                  (throw t))))]
+        (is (identical? failure
+                        (try (group/submit! g run-transaction op)
+                             (catch Throwable t t))))
+        (is (identical? failure @seen))
+        (is (= 1 @calls))
+        (is (= :recovered (group/submit! g #(% nil) (constantly :recovered))))))))
+
+(deftest failed-direct-writes-hand-off-to-queued-followers
+  (doseq [body-failure? [true false]]
+    (let [g (group/create 8)
+          entered (promise) release (promise)
+          failure (ex-info "direct write failed" {})
+          run-transaction (fn [execute]
+                            (let [result (execute nil)]
+                              (when (= :fail-commit result) (throw failure))
+                              result))
+          first-job (future
+                      (try
+                        (group/submit! g run-transaction
+                                       (fn [_]
+                                         (deliver entered true)
+                                         (assert (deref release 10000 false))
+                                         (when body-failure? (throw failure))
+                                         :fail-commit))
+                        (catch Throwable t t)))
+          jobs (atom [])]
+      (try
+        (is (deref entered 10000 false))
+        (reset! jobs (mapv (fn [n]
+                            (future (group/submit! g run-transaction (constantly n))))
+                          (range 7)))
+        (is (await! #(= 7 (queued g))))
+        (deliver release true)
+        (is (identical? failure (deref first-job 10000 ::timeout)))
+        (is (= (vec (range 7)) (mapv #(deref % 10000 ::timeout) @jobs)))
+        (finally
+          (deliver release true)
+          (doseq [job (cons first-job @jobs)] (deref job 10000 nil)))))))
+
 (deftest completed-followers-return-before-admission-is-released
   (let [acquisitions (AtomicInteger.)
         first-entered (promise) release-first (promise)
         commit-entered (promise) release-commit (promise)
         unlock-entered (promise) release-unlock (promise)
         lock (proxy [ReentrantLock] [true]
+               (tryLock []
+                 (if (let [^ReentrantLock this this] (proxy-super tryLock))
+                   (do (.incrementAndGet acquisitions) true)
+                   false))
                (lock []
                  (let [^ReentrantLock this this] (proxy-super lock))
                  (.incrementAndGet acquisitions))
@@ -37,15 +119,16 @@
         g (Group. lock (ConcurrentLinkedQueue.) 16 (AtomicBoolean. false))
         batches (atom [])
         returned (CountDownLatch. 8)
-        run-group (fn [^FastList requests]
-                    (swap! batches conj (.size requests))
-                    (let [results (group/execute requests nil)]
-                      (when (= 8 (.size requests))
-                        (deliver commit-entered true)
-                        (assert (deref release-commit 10000 false)))
-                      results))
+        run-transaction (fn [execute]
+                          (let [results (execute nil)
+                                n (batch-size results)]
+                            (swap! batches conj n)
+                            (when (= 8 n)
+                              (deliver commit-entered true)
+                              (assert (deref release-commit 10000 false)))
+                            results))
         first-job (future
-                    (group/submit! g run-group
+                    (group/submit! g run-transaction
                                    (fn [_]
                                      (deliver first-entered true)
                                      (assert (deref release-first 10000 false))
@@ -58,7 +141,7 @@
                       (future
                         (binding [*submitted-value* n]
                           (let [result (group/submit!
-                                         g run-group
+                                         g run-transaction
                                          (fn [_] [n *submitted-value*]))]
                             (.countDown returned)
                             result))))
@@ -89,20 +172,21 @@
         overlapping? (atom false)
         batches (atom [])
         calls (atom {})
-        run-group (fn [^FastList requests]
-                    (when-not (= 1 (.incrementAndGet executing))
-                      (reset! overlapping? true))
-                    (try
-                      (swap! batches conj (.size requests))
-                      (Thread/yield)
-                      (group/execute requests nil)
-                      (finally (.decrementAndGet executing))))]
+        run-transaction (fn [execute]
+                          (when-not (= 1 (.incrementAndGet executing))
+                            (reset! overlapping? true))
+                          (try
+                            (Thread/yield)
+                            (let [result (execute nil)]
+                              (swap! batches conj (batch-size result))
+                              result)
+                            (finally (.decrementAndGet executing))))]
     (dotimes [round 50]
       (let [start (promise)
             jobs (mapv (fn [n]
                          (future
                            @start
-                           (group/submit! g run-group
+                           (group/submit! g run-transaction
                                           (fn [_]
                                             (swap! calls update [round n] (fnil inc 0))
                                             (case (long (mod n 3)) 0 nil 1 false n)))))
@@ -121,19 +205,19 @@
         lock ^ReentrantLock (.-lock ^Group g)
         result (promise)
         done (CountDownLatch. 1)
-        run-group #(group/execute % nil)
+        run-transaction #(% nil)
         first-job (atom nil)
         follower (Thread.
                    (fn []
                      (try
                        (.interrupt (Thread/currentThread))
-                       (let [value (group/submit! g run-group (constantly :done))]
+                       (let [value (group/submit! g run-transaction (constantly :done))]
                          (deliver result [value (.isInterrupted (Thread/currentThread))]))
                        (catch Throwable t (deliver result t))
                        (finally (.countDown done)))))]
     (.lock lock)
     (try
-      (reset! first-job (future (group/submit! g run-group (constantly :first))))
+      (reset! first-job (future (group/submit! g run-transaction (constantly :first))))
       (is (await! #(.hasQueuedThreads lock)))
       (.start follower)
       (is (await! #(= 2 (queued g))))
@@ -149,10 +233,10 @@
         failure (ex-info "commit failed" {})
         commits (atom 0)
         calls (atom 0)
-        run-group (fn [requests]
-                    (group/execute requests nil)
-                    (swap! commits inc)
-                    (throw failure))
+        run-transaction (fn [execute]
+                          (execute nil)
+                          (swap! commits inc)
+                          (throw failure))
         jobs (atom [])]
     (.lock lock)
     (try
@@ -160,7 +244,7 @@
               (mapv (fn [_]
                       (future
                         (try
-                          (group/submit! g run-group (fn [_] (swap! calls inc)))
+                          (group/submit! g run-transaction (fn [_] (swap! calls inc)))
                           (catch Throwable t t))))
                     (range 8)))
       (is (await! #(= 8 (queued g))))
@@ -168,5 +252,5 @@
     (is (every? #(identical? failure %) (mapv #(deref % 10000 ::timeout) @jobs)))
     (is (= 8 @calls))
     (is (= 1 @commits))
-    (is (= :recovered (group/submit! g #(group/execute % nil)
+    (is (= :recovered (group/submit! g #(% nil)
                                    (constantly :recovered))))))